% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{bark}
\alias{bark}
\alias{hz}
\title{Barks and Hertz}
\usage{
bark()

hz()
}
\description{
\code{bark()} and \code{hz()} have been removed from \code{joeyr}. Please use the new package, \code{barktools},

\code{remotes::install_github("joeystanley/barktools")}

to access these functions. These will be phased out in future versions of \code{joeyr}.
}
