% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_transcriptions.R
\name{fave_to_wells}
\alias{fave_to_wells}
\title{Recode FAVE output to Wells.
Not as robust as `switch_allophones` but I'm tired of copying this code over.}
\usage{
fave_to_wells(.df)
}
\description{
Recode FAVE output to Wells.
Not as robust as `switch_allophones` but I'm tired of copying this code over.
}
