% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_grapes.R
\name{\%expanded_by\%}
\alias{\%expanded_by\%}
\title{Expand a range}
\usage{
range \%expanded_by\% expansion
}
\arguments{
\item{range}{A number. This number will be added to the highest and subtracted from the lowest values of the range}
}
\value{
A list of length 2
}
\description{
This function expands ranges by some value. Used in my dissertation code for axis ranges.
}
\examples{
c(1,5) \%expanded_by\% 1
}
