% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_tidy.R
\name{\%wi\%}
\alias{\%wi\%}
\title{Within}
\usage{
x \%wi\% range
}
\arguments{
\item{x}{A numeric vector.}

\item{range}{A numeric vector. The highest and lowest values will be used as the range.}
}
\value{
A logical vector.
}
\description{
Here's a very handy "within" range. Checks whether a number is within a range. Capable of handling lists so that it can apply to a whole column. Requires no additional packages.
}
\examples{
c(1:5) \%wi\% c(2,3)
}
