% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_filter.R
\name{joey_adjust_cutoff}
\alias{joey_adjust_cutoff}
\title{Adjust the cutoff score.}
\usage{
joey_adjust_cutoff(still_good, cutoff, boost, filter)
}
\description{
This function should only be used as called in \code{"joey_filter"}. This
adjusts the cutoff scores depending on the parameters of the previous
function.
}
