% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_filter.R
\name{joey_do_pca}
\alias{joey_do_pca}
\title{Run a Principal Components Analysis}
\usage{
joey_do_pca(still_good)
}
\description{
This function should only be used as called in \code{"joey_filter"}. It runs
a PCA on F1 and F2 of the data frame, saves the PCs, and returns the data
frame.
}
