% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_filter.R
\name{joey_filter}
\alias{joey_filter}
\title{Filter vowel measurements.}
\usage{
joey_filter(df, F1 = "F1", F2 = "F2", vowel = "vowel",
  method = "pca", filter = "cooksd", cutoff = 0, boost = 3,
  keep = FALSE)
}
\arguments{
\item{df}{A grouped dataframe including F1, F2, and vowel columns.}

\item{F1}{A quoted string of the column name containing F1 measurements. By
default, a column named "F1" is used.}

\item{F2}{A quoted string of the column name containing F2 measurements. By
default, a column named "F2" is used.}

\item{vowel}{A quoted string of the column name containing the "vowel" column.
This would be the column that identifies what vowel quality the observations
in the row belong to. By default, a column named "vowel" is used.}

\item{method}{A string. By default, \code{"pca"} is used, which will perform a
 principal components analysis on the data and do a regression analysis on
 the first two principal components. This is prefered since it takes into
 account the correlation in the data.

 As an alternative, \code{"f1f2"} can be used instead (technically any string
 other than \code{"pca"}), which does a regression on the raw formant
 measurements. This is akin to filtering by normalized F1 and F2 (perhaps by
 removing anything more than 3 standard deviations from the mean), but it
 finds a new mean and standard deviation after removing each point.}

\item{filter}{A string. By default, the method is "cooksd", meaning that
outliers are detected using Cook's D. Any observation with a Cook's D
greater than the \code{"cutoff"} is considered an outlier. If any other
string is used, such as \"z-score\" than outliers are detected based on a
z-score. The data is normalized and any observation further than *n*
standard deviations from the mean is considered an outlier. The cutoff, in
standard deviation units, is specified in the \code{"cutoff"} parameter.}

\item{cutoff}{A number. When using Cook's D as the \code{"filter"}, any
 observation with a Cook's D higher than this value is considered an outlier.
 By default, the rule of thumb for Cook's D is used: \deqn{4/(n - k - 1)}
 where \emph{n} is the number of observations, and \emph{k} is the number of
 variables, which in this case, is 2 (PC1 and PC2 or F1 and F2).

 Usually, the number of observations will be different for each vowel within
 the same speaker, so this value will be different from vowel to vowel. If,
 however, a specific cutoff value is wanted that should be used for all
 vowels regardless of how many observations there are or their distribution,
 it can be specified here.

 When using z-score as the \code{"filter"}, the default is that any
 observations further than 2 standard deviations from the mean are considered
 outliers. This can changed by specifying a number here.}

\item{boost}{A number. When using \code{"Cook's D"} as the \code{"method"},
the rule of thumb cutoff value is often too strong and removes too many
observations. To make the filter weaker, the cutoff value is multiplied by
the \code{"boost"} value. By default, it is set to \code{"3"}. To apply no
filter and to use the exact cutoff value, set this value to \code{"1"}. If
the \code{"cutoff"} value is specified or if z-score is used as the filter,
this parameter is quiety ignored.}

\item{keep}{Logical. By default, outliers are removed from the data frame.
However, if you would like to keep them (for examination), set this argument
to \code{"TRUE"}. This will create a new TRUE/FALSE column called
\code{"is_outlier"} that will identify whether that particular observation
was determined to be an outlier.}
}
\value{
A dataframe. Outliers are removed if \code{"keep = FALSE"} so there
 will be fewer rows. If \code{"keep = TRUE"}, a new variable,
 \code{"is_outlier"} will be tagged on the end.
}
\description{
Description goes here.
}
\examples{
laterals <- read.csv("joey.csv") \%>\%
    group_by(vowel)

# Detect outliers using Cook's D based on a PCA regression.
joey_filter(laterals)
# a stronger filter
joey_filter(laterals, boost = 2)
# with blanket cutoff score
joey_filter(laterals, cutoff = 0.1)

# Detect outliers using Cook's D based on F1 and F2 values.
joey_filter(laterals, method = "f1f2")

# Detect outliers using z-score based on PCA regression.
joey_filter(laterals, filter = "z-score")

# Detect outliers using z-score based on F1 and F2 values.
joey_filter(laterals, method = "f1f2", filter = "z-score")
# Based z > 2.5 instead of z > 2.
joey_filter(laterals, method = "f1f2", filter = "z-score", cutoff = 2.5)

# Specify column names.
joey_filter(laterals, F1 = "F1.50.", F2 = "F2.50.", vowel = "phoneme")

# Keep the outliers.
joey_filter(laterals, keep = TRUE)
}
