% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joey_formants}
\alias{joey_formants}
\title{Joey's formant measurements}
\format{
An object of class \code{data.frame} with 3504 rows and 43 columns.
}
\usage{
joey_formants
}
\description{
A dataset containing 3,504 formant measurements. After reading 300 sentences,
I used DARLA to automatically transcribe, force-align, and extract formants
from the audio. Not the cleanest dataset (plus, I had a cold that day), but 
it is a decent enough sample of one speaker's American English vowels.
}
\details{
The data has been unaltered from the output FAVE produced.
}
\keyword{datasets}
