% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_filter.R
\name{joey_rm_outliers}
\alias{joey_rm_outliers}
\title{Remove outliers}
\usage{
joey_rm_outliers(df, this_combo, F1, F2, vowel, method, filter, cutoff,
  boost, keep)
}
\description{
This function should only be used as called in \code{"joey_filter"}. This is actually the function that does all the work and contains the loop but you can only use it by calling the \code{"joey_filter"} function.
}
