% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{lbms_index}
\alias{lbms_index}
\title{Low-Back-Merger Shift Index}
\usage{
lbms_index(df, vowel_col, F1_col, F2_col, beet, bit, bet, bat)
}
\arguments{
\item{df}{The dataframe containing the formant measurements you want to base the calculation off of.}

\item{vowel_col}{The name of the column containing the name of the vowel (e.g. \code{vowel})}

\item{F1_col}{The name of the column containing the F1 measurements (e.g. \code{F1_norm})}

\item{F2_col}{The name of the column containing the F2 measurements (e.g. \code{F2_norm})}

\item{beet}{A string that indicates which vowels belong to the BEET class of words (e.g. \code{"IY"})}

\item{bit}{A string that indicates which vowels belong to the BIT class of words (e.g. \code{"IH"})}

\item{bet}{A string that indicates which vowels belong to the BET class of words (e.g. \code{"EH"})}

\item{bat}{A string that indicates which vowels belong to the BAT class of words (e.g. \code{"AE"})}
}
\value{
A \code{summarize}d dataframe with the LBMS Index.
}
\description{
A function to quickly calculate the Low-Back-Merger Shift Index, per Becker (2019).
}
\details{
If you would like to calculate the LBMS Index for each speaker, be sure to 
the data beforehand with \code{group_by()} (see the examples). Also, per Becker's
(2019) recommendation, it is recommended that you exclude tokens before nasals, 
laterals, rhotics, and /g/ before the calculation.

Note that this is a new function and has not been tested robustly yet.
}
\examples{
library(tidyverse)
data(joey_formants) # Not the best example because there's only one speaker.
joey_formants \%>\%
   filter(!fol_seg \%in\% c("L", "R", "N", "M", "NG")) \%>\%
   group_by(name) \%>\%
   lbms_index(vowel, F1_LobanovNormed_unscaled, F2_LobanovNormed_unscaled, 
              "IY", "IH", "EH", "AE")

}
\references{
Becker, Kara, ed. The Low-Back-Merger Shift: Uniting the Canadian Vowel Shift, 
the California Vowel Shift, and Short Front Vowel Shifts across North America. 
Publication of the American Dialect Society 104. Durham, NC: Duke University 
Press, 2019.
}
