% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_tidy.R
\name{move_x_after_y}
\alias{move_x_after_y}
\title{Move one column after another}
\usage{
move_x_after_y(df, colx, coly)
}
\arguments{
\item{df}{A data frame.}

\item{colx}{A column in that dataframe.}

\item{coly}{Another column in that dataframe.}
}
\value{
A dataframe, with colx immediately after coly.
}
\description{
This quite simply moves one column after another in a data frame. Just a simple way to rearrange a couple columns without having to use select. It saves a bit of typing and makes it more clear what it is that's happening.
}
\examples{
sample_df <- data.frame(first = 1:3, middle = 4:6, last = 7:9)
move_x_after_y(sample_df, last, first)
}
