% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{norm_anae}
\alias{norm_anae}
\title{ANAE Vowel Normalization}
\usage{
norm_anae(df, hz_cols, vowel_id, speaker)
}
\arguments{
\item{df}{The dataframe containing the formant measurements you want to normalize.}

\item{hz_cols}{A list of columns (unquoted) containing the formant measurements themselves.}

\item{vowel_id}{The name of the column containing unique identifiers per vowel.}

\item{speaker}{The name of the column containing unique identifiers per speaker (usually the column containing the speaker name).}
}
\value{
The same dataframe, but with new column(s), suffixed with "_anae" that have the normalized data.
}
\description{
This is a a tidyverse-compatible function that makes it easy to normalize 
your data using the method described in the Atlas of North American English. 
NOTE: This is brand new and has not been tested robustly.
}
\details{
It is unclear how the ANAE function should work with trajectory data.
This function pools all data together and normalizes it together, which means
one small modification was required to calculate the G value: I had to add
the average number of time points per vowel token in the denominator. Not sure
if that's how it should be done, but it makes sense to me and returns sensible
results.
}
\examples{
library(tidyverse)
# Not the best example because there's only one speaker.
data(joey_formants)
joey_formants \%>\%
   rowid_to_column("token_id") \%>\%
   norm_anae(hz_cols = c(F1, F2), vowel_id = token_id, speaker = name)
}
