% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{pillai}
\alias{pillai}
\title{Calculate Pillai scores}
\usage{
pillai(...)
}
\description{
This is just a shortcut to run an MANOVA and return just the pillai score.
}
\examples{
pillai(cbind(F1, F2) ~ vowel + fol_seg, data = df)

}
