% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_tidy.R
\name{spread_n}
\alias{spread_n}
\title{spread() on multiple values.}
\usage{
spread_n(df, key, value)
}
\description{
Use tidyr::spread but on multiple columns. I haven't been able to get this function to work well when there are two numeric columns I need to work with. All credit goes to "danr" on the RStudio Forums who answered my quesiton [here](https://community.rstudio.com/t/spread-with-multiple-value-columns/5378/2).
}
\examples{
# Create a data frame with some sample formants. Crucially there are two numeric values for each row.
df1 <- data.frame(word = rep(c("dog", "cat", "fish"), 2),
formant = rep(c("F1", "F2"), each = 3),
hz = c(490, 500, 510, 1590, 1600, 1610),
hz_bark = bark(c(490, 500, 510, 1590, 1600, 1610)))
# Now spread them with each formant and type of measurement in its own column.
df1 \%>\%
   spread_n(formant, c(hz, hz_bark))

df2 <- data.frame(speaker = rep(c("Amy", "Bob"), each = 3),
                  word = rep(c("dog", "cat", "fish"), 2),
                  F1 = c(490, 500, 510, 390, 400, 410),
                  F2 = bark(c(1590, 1600, 1610, 1390, 1400, 1410)))
df2 \%>\%
    spread_n(word, c(F1, F2))
df \%>\%
   spread_n(speaker, c(F1, F2))
}
