% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ucfirst}
\alias{ucfirst}
\title{Capitalize the first letter}
\usage{
ucfirst(str)
}
\arguments{
\item{str}{A string.}
}
\value{
The same string with the first letter capitalized.
}
\description{
Capitalizes the first character is a string. Not particularly robust. Copied over from Perl code.
}
\examples{
ucfirst("foo bar") # "Foo bar"
}
