% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{bark}
\alias{bark}
\alias{hz}
\title{Barks and Hertz}
\usage{
bark(f)

hz(z)
}
\arguments{
\item{x}{A number}
}
\value{
The converted number.
}
\description{
Convert Hz into Barks and vice versa. The formula and inverse are taken from Traunmüller (1990).
}
\examples{
bark(1000) # 8.527432
hz(8.527432) # 1000
}
