% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{eucl_dist}
\alias{eucl_dist}
\title{Measure euclidean distance}
\usage{
eucl_dist(x1, x2, y1, y2)
}
\arguments{
\item{x1, x2, y1, y2}{A number}
}
\value{
The distance between the two points (a number).
}
\description{
This function takes the x and y coordinates of two points and finds
the euclidean distance between the two.
}
\examples{
eucl_dist(1,2,1,2)
}
