% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociophonetics.R
\name{norm_deltaF}
\alias{norm_deltaF}
\title{∆F Normalization}
\usage{
norm_deltaF(df, .F1, .F2, .F3, .F4)
}
\arguments{
\item{df}{The data frame containing the formant measurements you want to 
normalize}

\item{.F1, .F2, .F3, .F4}{The (unquoted) name of the column containing the F1 
measurements. The first three are required, but you may leave off F4.}
}
\value{
The original dataframe with new columns (suffixed with 
\code{_deltaF}) containing the normalized measurements.
}
\description{
Normalize vowel formant measurements with ∆F (see Johnson 2020). This 
function is intended to be used within a tidyverse pipeline.
}
\details{
You will need to group the data by speaker with \code{group_by()} before 
applying this function if you want to normalize the data by speaker.

The data must be numeric, and there cannot be any NAs. So, if you're using 
data extracted from Praat, you may have to filter out bad F3 and F4 data
and then convert the column to numeric. See the example.

Note that this is a new function and has not been tested robustly yet.
}
\examples{
library(tidyverse)
# Not the best example because there's only one speaker.
data(joey_formants) 

# Clean the data first
joey_formants_clean <- joey_formants \%>\%
   # Remove tokens without an F3 or F4 measurement
   filter(F3 != "None") \%>\%
   # Convert F3 to numeric
   mutate(F3 = as.numeric(F3))
   
 # Run the function
 joey_formants_clean \%>\%  
   group_by(name) \%>\%
   norm_deltaF(F1, F2, F3)
   
}
\references{
Johnson, Keith. The ΔF Method of Vocal Tract Length Normalization for Vowels. 
Laboratory Phonology: Journal of the Association for Laboratory Phonology 11, 
no. 1 (July 22, 2020): 10. https://doi.org/10.5334/labphon.196.
}
