% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joeyr_math.R
\name{tidy_mahalanobis}
\alias{tidy_mahalanobis}
\title{Calculate Mahalanobis Distance}
\usage{
tidy_mahalanobis(...)
}
\arguments{
\item{...}{Names of columns that should be included in the Mahalnobis distance. For vowel data, this is typically your F1 and F2 columns.}
}
\value{
A vector that contains the Mahalabis distances for each observation.
}
\description{
This is a tidyverse-tidyverse-compatible version of the `mahalanobis` function. It just 
makes it easier to include it as part of a `dplyr::mutate`. One small modification that 
`mahalnobis` function does not do is that if there are fewer than 5 measurements in the
dataset, , it returns them all as having a distance of zero. Prevents some errors that 
way. Requires the MASS package to be installed, but this this not necessarily loaded.
}
\examples{
require(dplyr)

# Create some artificial vowel data
tibble(F1 = rnorm(50, 0, 1),
       F2 = rnorm(50, 0, 1)) \%>\%
 # Calculate the Mahalanobis distance
 mutate(mahal_dist = tidy_mahalanobis(F1, F2))

# Same thing, but using base R: 
vowel_data <- data.frame(F1 = rnorm(50, 0, 1),
                         F2 = rnorm(50, 0, 1)) 
vowel_data$mahal_dist <- tidy_mahalanobis(vowel_data$F1, vowel_data$F2)
head(vowel_data)
}
